backend/requirements.txt
celery==4.3.0
Django==2.2
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-environ==0.4.5
django-extensions==2.1.4
django-filebrowser-no-grappelli==3.7.7
django-filter==2.1.0
django-jsonfield==1.0.1
django-redis==4.10.0
django-tinymce4-lite==1.7.4
djangorestframework==3.9.2
djangorestframework-datatables
djangorestframework_simplejwt==4.1.5
flower==0.9.3
ipython==7.3.0
Pillow==6.0.0
mysqlclient==1.4.1
python-memcached===1.59
redis==3.2.1
uwsgi==2.0.18
sentry-sdk==0.7.10
